requirements.txt
attrs
arrow
requests
ujson
